Move unstructured_api.py to v2 code
